Dzien2/09.ReadMeIfYouCan-main/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.8.0

# Azure services
azure-identity>=1.16.0
azure-ai-documentintelligence>=1.0.0

# OpenAI integration
openai>=1.70.0
tiktoken>=0.7.0; platform_system != "Windows" or platform_machine != "ARM64"

# PDF processing
pdf2image>=1.17.0
Pillow>=10.2.0
pypdfium2>=4.30.0

# Date handling
python-dateutil>=2.8.2

